requirements.txt
tensorflow==2.16.1
numpy>=1.23.5,<2.0
pandas>=2.2,<3
scikit-learn>=1.4,<2
matplotlib>=3.8,<4
pillow>=10,<12
opencv-python<4.11
kagglehub>=0.3
tqdm>=4.67
notebook>=7
ipykernel>=6
streamlit>=1.34,<2
